gui: Hook up MQTT backend to the GUI
The toolbar now has a new button: "Connect". Currently, clicking it
makes the backend connect to a preset server. Ideally, it will spawn
a dialog window where the user can specify information about the server.

Another new thing is a status bar at the bottom of the window. It seemed
like an ideal place for showing the state of the connection. Currently
it shows either "connected" or "disconnected". In the future that could
be a diode of some sort, name of the server (broker) and more.