explorer: Populate TreeView with topics and wire it to MainWindow
This gives ExplorerPage the ability to add items to the TreeView
representing topics of messages.

Major part of this is the addition of a new class: MessageStore. It's
purpose is to aggregate received messages and provide handy methods for
interacting with the store. Currently, it is only good for grouping new
messages into batches. This is very important for keeping the GUI
responsive.