explorerpage: Add means to view message list in a topic
With this the right part of the Explorer widget finally gets some
functionality. When a topic is clicked, the message list is populated
with all messages related to the said topic. When one of the items in
the list is further clicked on, the message preview is populated with
the content of the message.

TODO:
- elipsize messages in the message history list
- add colors to the messages