explorerpage: Reverse message history list, truncate previews
It seemed more natural for the message history list to grow down instead
of growing up. This required putting in place a hack that prevents
having to rething the way messages are stored.

Messages are now truncated to 26 characters as per instructions. To view
the whole message, click on it in the list and the preview window will
be populated. In the future and additional way of handling e.g. images
will have to be added.

Foundation for recognizing images in payloads has been laid but it does
not function, yet. I spent significant time trying to make it work but
it is no longer worth it. I might return to it later on.