mainWindow: Unify toolbars and make it persistent
Having two toolbars for the same thing is unnecessary.

A toolbar is a widget that can be hidden. Normally, a toolbar can be
brought back by: right clicking on another toolbar, right clicking on a
menu or having an action in another toolbar/menu. I do not want to deal
with this, so I connected the changeVisibility method to show() which
effectively makes it impossible to hide the toolbar.