messagestore: Add action for saving snapshot of topics + latest payloads
Required by the assignment.

This adds a new action into the toolbar that is visible only in Explorer
mode. Upon clicking it shows a file dialog for choosing a directory.
After choosing a valid directory a snapshot of the current state of
topics is written in the specified directory. The topics are arranged
in the same way they'd be in the tree view excluding only the two root
nodes (server node + root node). For each topic the latest message is
saved.