mqttmanager: Adjust connection options and default subscription
It makes more sense to start with a clean plate every time a client
connects to the server. Also, it is good to let the library handle
automatic reconnection in case the app loses it.

Due to the design of the app (Explorer for watching every topic known to
the broker and Dashboard for nice view of specific topics) it makes more
sense to subscribe to _all_ topics.

Also, no longer close the app when failing to connect to a server.