messagestore: Change the way messages are stored
Until now messages in a topic were stored in QList<QString>. The use of
QString caused mangling of binary data and thus the inability to
recognize them as proper images. Also, there was no information about
the time of arrival of an message and if it was sent or received from
a broker.

The introduced Message class contains: the payload, time of arrival and
the type of the message.

Time of arrival and type of the message is further used when filtering
received messages right after sending a message to the broker. This is
needed to satisfy a requirement set by the faculty to effectively
discern between sent and received messages.