Add simulator
A simulator of traffic is one of the requirements for the assignment.

This adds a new page/stack to the main window. The window has two
actions related to it: load configuration, toggle simulation.

A simulation is very basic and only supports sending payloads in a set
interval.

The configuration is done via a JSON file with a specific format. More
information can be found in the project's README file.